v1.1 Recursiveness is a Luxury you can Afford
Added a new feature that now allows the creation of trade offers larger than 1m tons at the same price without having to return to the market page first.
This feature also disables the buy/sell button for 5s when active to remove the 5s delay between the creation of trade offers.